financial-solutions/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
lazy-object-proxy==1.4.2
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typed-ast==1.4.0
typing-extensions==3.10.0.0
